requirements_temp.txt
antlr4-python3-runtime==4.9.3
Brotli==1.0.9
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.16.1
fsspec==2024.10.0
h5py==3.11.0
hydra-core==1.2.0
idna==3.7
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.1
numpy==1.24.3
omegaconf==2.3.0
packaging==24.1
pillow==10.4.0
pip==24.2
platformdirs==3.10.0
pointnet2_ops==3.0.0
pooch==1.7.0
PySocks==1.7.1
pytorch-triton==0.0.1
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.3.0
scipy==1.10.1
setuptools==75.1.0
sympy==1.13.3
tabulate==0.9.0
threadpoolctl==3.5.0
torch==2.1.2+cu118
torchaudio==0.13.1+cu117
torchvision==0.16.2+cu118
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
wheel==0.44.0
win-inet-pton==1.1.0
zipp==3.20.2
